+++ packages/styleguide/src/data/components/Checkbox.ts
@@ -1,4 +1,4 @@
-import { ComponentData } from '../data';
+import { ComponentData } from "../types";
 
 const Checkbox: ComponentData = {
     description: `
